#6 add redux, router into modules
