+++ services/employee-info/src/app/providers/index.ts
@@ -1,0 +1,5 @@
+import { compose } from '@modules/info/shared'
+import { withRouter } from './withRouter'
+import { withStore } from './withStore'
+
+export const withProviders = compose(withRouter, withStore)
