+++ services/employee-info/src/index.js
@@ -1,1 +1,1 @@
-import('./app')
+import('.')
