+++ services/employee-info/tsconfig.json
@@ -1,7 +1,10 @@
 {
   "extends": "../../tsconfig.json",
-  "include": [
-    "src"
-  ],
+  "compilerOptions": {
+    "baseUrl": "src",
+    "paths": {
+      "@modules/info/*": ["*"]
+    }
+  }
 
 }
