+++ services/employee-time-off/src/app/router/index.tsx
@@ -1,0 +1,7 @@
+import { Routes, Route } from 'react-router-dom'
+
+export const Pages = () => (
+  <Routes>
+    <Route path='/time-off' element={<div>Hello time-off</div>} />
+  </Routes>
+)
