+++ services/employee-time-off/webpack.config.js
@@ -80,8 +80,8 @@ module.exports = {
         './TimeOff': './src/app',
       },
       shared: {
-        react: {
-          requiredVersion: deps.react,
+        'react': {
+          requiredVersion: deps['react'],
           singleton: true,
         },
         'react-dom': {
