+++ services/employee-vacation/src/app/providers/withRouter.tsx
@@ -1,0 +1,3 @@
+import { BrowserRouter } from 'react-router-dom'
+
+export const withRouter = (component: () => React.ReactNode) => () => <BrowserRouter>{component()}</BrowserRouter>
