+++ services/employee-vacation/src/app/providers/withStore.tsx
@@ -1,0 +1,4 @@
+import { Provider } from 'react-redux'
+import { store } from '../store'
+
+export const withStore = (component: () => React.ReactNode) => () => <Provider store={store}>{component()}</Provider>
