+++ services/employee-vacation/src/app/router/index.tsx
@@ -1,0 +1,7 @@
+import { Routes, Route } from 'react-router-dom'
+
+export const Pages = () => (
+  <Routes>
+    <Route path='/vacation' element={<div>Hello vacation</div>} />
+  </Routes>
+)
