+++ services/employee-vacation/tsconfig.json
@@ -1,7 +1,10 @@
 {
   "extends": "../../tsconfig.json",
-  "include": [
-    "src"
-  ],
+  "compilerOptions": {
+    "baseUrl": "src",
+    "paths": {
+      "@modules/vacation/*": ["*"]
+    }
+  }
 
 }
