+++ services/entry-point/src/app/providers/index.ts
@@ -1,0 +1,3 @@
+import { withRouter } from './withRouter'
+
+export const withProviders = withRouter
