+++ services/entry-point/src/index.tsx
@@ -1,0 +1,5 @@
+import ReactDOM from 'react-dom/client'
+import App from './app'
+
+const root = ReactDOM.createRoot(document.getElementById('root')!)
+root.render(<App />)
